Reviewing and concatenating the last 2 paragraphs about the process to request inclusion and removal from contributors' list
The 2 last paragraphs provide details about the process to ask inclusion or removal from the contributor's list.
My suggestion is to concatenate the 2 paragraphs and the instructions about who is eligible and then how to ask inclusion and removal from list.
I understand the message about this process is friendly and flexible, as the request can be done by using GitHub features (Pull-Request/Issues), by mailing list, or by contacting the maintainer. However, I think it can be better to unify and make it simpler, and also clarify that the person should open a pull-request (and not a new issue) for either ask inclusion or removal.